requirements.txt
flask
psycopg2-binary
pandas
xlsxwriter
gunicorn
